requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
Django==3.0.6
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-ckeditor==5.9.0
django-cors-headers==3.3.0
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
future==0.18.2
idna==2.9
kombu==4.6.11
markdown2==2.3.9
Pillow==7.1.2
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
